aiagent/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.3
pyjwt==2.8.0
httpx==0.27.0